Handles non-alphanumeric user ids
In messenger chat, when a user's id has a character that would cause an
html attribute to be invalid, anyone chatting with that user will not be
able to see any messages. This occurs because the userid is used to
create the id for an element that needs to be retrieved to add messages
inside.